use multiple swap-in addresses
We create and monitor a fixed number of different swap-in addresses.
Peer will choose an empty swap-in addresses if there is one, or a random address if there are none.
We limit the number of monitored swap-in addresses to avoid hitting resource limits on Electrum servers.
A few hundred addresses (100 or 200) is a safe limit.