Use multiple swap-in addresses, and select an unused address if possible
We create and monitor a fixed number of different swap-in addresses.
Peer will choose an empty swapin addresses if there is one, or a random address if there are none.
We limit the number of monitored swap-in addresses to avoid hitting resource limits on Electrum servers.
A few hundred addresses (100 or 200) is a safe limit.