Implement the option_simple_close protocol
We update the `Negotiating` state to exchange the `closing_complete` and
`closing_sig` messages, and allow RBF-ing previous transactions and
updating our closing script.

We stay in that state until one of the transactions confirms, or a force
close is detected. This is important to ensure we're able to correctly
reconnect and negotiate RBF candidates.